short to int, float to double

+++ Heuristique/PFE/Data.cpp
@@ -18,59 +18,59 @@ int M() // Returns the number of machines
 { return Data.NbMachines; }
 
 // For tasks
-short int nc(unsigned int i) // Returns the required CPU load of task i
+int nc(unsigned int i) // Returns the required CPU load of task i
 { return Data.ListOfTasks[i].QtyCPU ;}
-short int ng(unsigned int i) // Returns the required GPU load of task i
+int ng(unsigned int i) // Returns the required GPU load of task i
 { return Data.ListOfTasks[i].QtyGPU ;}
-short int nr(unsigned int i) // Returns the required RAM of task i
+int nr(unsigned int i) // Returns the required RAM of task i
 { return Data.ListOfTasks[i].QtyRAM ;}
-short int nh(unsigned int i) // Returns the required harddrive of task i
+int nh(unsigned int i) // Returns the required harddrive of task i
 { return Data.ListOfTasks[i].QtyHDD ;}
-short int u(unsigned int i, unsigned int t) // Returns if task i is likely to be processed at time t
+int u(unsigned int i, unsigned int t) // Returns if task i is likely to be processed at time t
 { return Data.ListOfTasks[i].LIsToBeProcessed[t] ;}
-short int a(unsigned int i, unsigned int k) // Returns 1 if tasks i and k have an affinity; 0 otherwise
+int a(unsigned int i, unsigned int k) // Returns 1 if tasks i and k have an affinity; 0 otherwise
 { return Data.ListOfTasks[i].LAffinities[k];}
-short int q(unsigned int i, unsigned int j) // Returns 1 if task i can be assigned to machine j; 0 otherwise
+int q(unsigned int i, unsigned int j) // Returns 1 if task i can be assigned to machine j; 0 otherwise
 { return Data.ListOfTasks[i].LPreAssignement[j] ;}
-short int b(unsigned int i, unsigned int j) // Returns the bandwidth required by tasks i and j to communicate over the network
+int b(unsigned int i, unsigned int j) // Returns the bandwidth required by tasks i and j to communicate over the network
 { return i<=j? Data.Network.ComBandwidth[i][j]:Data.Network.ComBandwidth[j][i] ;}
-short int R(unsigned int i) // Returns 1 if task i is preemptable; 0 otherwise
+int R(unsigned int i) // Returns 1 if task i is preemptable; 0 otherwise
 { return Data.ListOfTasks[i].isPreemptable ;}
-short int rho(unsigned int i) // Returns the unitary penalty induced by preempting taks i
+int rho(unsigned int i) // Returns the unitary penalty induced by preempting taks i
 { return Data.ListOfTasks[i].CostPreemption ;}
-short int mt(unsigned int i) // Returns the migration time of task i
-{ return (short int)ceil((float)(nh(i)+nr(i))/(float)b(i,i));}
-short int rt(unsigned int i, unsigned int j) //Returns the resuming time of task i on machine j
-{  return (short int)ceil((float)nr(i)/(float)v(j));}
+int mt(unsigned int i) // Returns the migration time of task i
+{ return (int)ceil((double)(nh(i)+nr(i))/(double)b(i,i));}
+int rt(unsigned int i, unsigned int j) //Returns the resuming time of task i on machine j
+{  return (int)ceil((double)nr(i)/(double)v(j));}
 
 // For machines
-short int mc(unsigned int j) // Returns the maximum accepted CPU load of machine j
+int mc(unsigned int j) // Returns the maximum accepted CPU load of machine j
 { return  Data.ListOfMachines[j].QtyCPU ;}
-short int mg(unsigned int j) // Returns the maximum accepted GPU load of machine j
+int mg(unsigned int j) // Returns the maximum accepted GPU load of machine j
 { return  Data.ListOfMachines[j].QtyGPU ;}
-short int mr(unsigned int j) // Returns the maximum accepted RAM load of machine j
+int mr(unsigned int j) // Returns the maximum accepted RAM load of machine j
 { return  Data.ListOfMachines[j].QtyRAM ;}
-short int mh(unsigned int j) // Returns the maximum accepted harddrive load of machine j
+int mh(unsigned int j) // Returns the maximum accepted harddrive load of machine j
 { return  Data.ListOfMachines[j].QtyHDD ;}
-short int alphac(unsigned int j) // Returns the cost for using CPU of machine j
+int alphac(unsigned int j) // Returns the cost for using CPU of machine j
 { return  Data.ListOfMachines[j].CostCPU ;}
-short int alphag(unsigned int j) // Returns the cost for using GPU of machine j
+int alphag(unsigned int j) // Returns the cost for using GPU of machine j
 { return  Data.ListOfMachines[j].CostGPU ;}
-short int alphar(unsigned int j) // Returns the cost for using RAM of machine j
+int alphar(unsigned int j) // Returns the cost for using RAM of machine j
 { return  Data.ListOfMachines[j].CostRAM ;}
-short int alphah(unsigned int j) // Returns the cost for using Harddrive of machine j
+int alphah(unsigned int j) // Returns the cost for using Harddrive of machine j
 { return  Data.ListOfMachines[j].CostHDD ;}
-short int beta(unsigned int t) // Returns the cost for turning on a machine at time t
+int beta(unsigned int t) // Returns the cost for turning on a machine at time t
 { return  Data.CostTurnOn[t] ;}
-short int v(unsigned int j) // Returns the speed of machine j in loading a processing context
+int v(unsigned int j) // Returns the speed of machine j in loading a processing context
 { return  Data.ListOfMachines[j].SpeedContext ;}
 
 // For the network
-short int maxb() // Returns the maximum bandwidth of any edge
+int maxb() // Returns the maximum bandwidth of any edge
 { return  Data.Network.MaxBandwidth ;}
-short int NbEdges() // Returns the number of Edges in the network
+int NbEdges() // Returns the number of Edges in the network
 { return  Data.Network.NbEdges ;}
-short int NbMachEdge(unsigned int e) // Returns the number of couples of machines that use the edge e
+int NbMachEdge(unsigned int e) // Returns the number of couples of machines that use the edge e
 { return  Data.Network.NbMachinesByEdge[e];}
 void CoupleMachines(unsigned int e, unsigned int pos, unsigned int &Mach1, unsigned int &Mach2) // Mach1 and Mach2 are the machines in position pos in the list of couples usinge the edge e
 {
